Multiple patches from jcleaver: - network tests resulting in odd connect()-errors now include the test definiti$ - r6905 from trunk: fix RRD update of xymongen status (blanks must be removed) - sendmsg() debug code had syntax error - stackio: Must initialise eolchar - xymond: Timers for statistics update and purple-check changed from hardcoded $ - Clients now use "xymon --version" as the default for client-version data - xymoncmd: Use $HOSTNAME as the hostname, if defined, before running "uname". $ - xymoncmd: Fix buffer overflow in defining SERVEROSTYPE environment variable. - xymonnet and xymongen can now run via CRONDATE definitions. - xymonlaunch: Search for tasks.cfg in /etc, /etc/xymon, /etc/xymon-client and $ - xymond_client: Permit data from unknown client types, but parse only the "mes$ - Linux client: Send error-output from netstat/ifconfig to /dev/null - loadhosts: Handle "vpage", "vsubpage" and "vsubparent" correctly. - loadhosts: Increase default hosts.cfg filebuffer size to 20 kB - xymond: Send "drop" and other admin-commands to the client-change and user-ch$ - xymond_channel: Make the timeout for a worker to handle a message configurabl$ - xymond_channel: Don't commit genocide on xymonlaunch when shutting down in ne$ - xymond_rrd: Add --no-rrd option to let all graph data flow to an external pro$ - Linux client: Format ps output to align columns better. - Add SHELL variable to xymonserver.cfg, and use it when we need to run an exte$ - "msgs" status: Change text from "System logs" to "Log files" - "xymongen" and "xymond" status: Standardize format of first status line.
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7205 44351d6e-118b-4698-b696-ce33095ecaa4

+++ trunk/lib/readmib.h
@@ -68,7 +68,7 @@ typedef struct mibidx_t {
 	enum mibidxtype_t idxtype;		/* How to interpret the key */
 	char *keyoid;				/* Key OID */
 	void *rootoid;				/* Binary representation of keyoid */
-	unsigned int rootoidlen;		/* Length of binary keyoid */
+	size_t rootoidlen;			/* Length of binary keyoid */
 	struct mibidx_t *next;
 } mibidx_t;
